+++ test/github-issues/3857/entity/Men.ts
@@ -1,0 +1,9 @@
+import {Person} from "./Person";
+import {ChildEntity, Column} from "../../../../src";
+
+@ChildEntity()
+export class Men extends Person {
+
+    @Column("varchar")
+    beardColor: string;
+}