+++ test/github-issues/3857/entity/Women.ts
@@ -1,0 +1,9 @@
+import {Person} from "./Person";
+import {ChildEntity, Column} from "../../../../src";
+
+@ChildEntity()
+export class Women extends Person {
+
+    @Column("int")
+    brassiereSize: number;
+}